Move the colmena thinker inside the module
